Add tmp dir arg to circle_finder...sh
Here is one way to go about adding the path as an argument to the affected script. I haven't tested this. It should fail if no eighth argument is supplied, or if the directory supplied does not exist.

You would also need to modify the sbatch script that calls this one in order for this to work. Please check the ticket we had recently for a way to do this.